+++ gateway/src/main/java/org/acumos/federation/gateway/Application.java
@@ -3,6 +3,7 @@
  * Acumos
  * ===================================================================================
  * Copyright (C) 2017-2019 AT&T Intellectual Property & Tech Mahindra. All rights reserved.
+ * Modifications Copyright (C) 2020 Nordix Foundation.
  * ===================================================================================
  * This Acumos software file is distributed by AT&T and Tech Mahindra
  * under the Apache License, Version 2.0 (the "License");
@@ -203,6 +204,10 @@ public class Application {
 		return new LogstashServiceImpl();
 	}
 
+	@Bean
+	DeployedModelService deployedModelService() {
+		return new DeployedModelServiceImpl();
+	}
 
 	static Docket getApi() {
 		String version = Application.class.getPackage().getImplementationVersion();